Add files via upload
JavaScript
Create a loop the iterates up to 100 while outputting “Anna” at multiples of 3, “belle” at multiples of 5, and “Annabelle” at multiples of 3 and 5.	(5 points)
Create a JavaScript function that returns true if two words are anagrams e.g. “Elbow” and “Below” are anagrams					(5 points)

SQL
Your are tasked with creating a database model to capture university data for College X
Create an ER diagram that captures the relationship between a Student, a Course, and a Lecturer					(10 points)
Write SQL queries to create the entities mentioned in 3.a	(5 points)
Write an SQL query that outputs all students that are taught by a specific Lecturer							(5 points)